+++ schema/src/main/scala/org/constellation/schema/package.scala
@@ -31,8 +31,9 @@ package object schema {
 
   }
 
-  def hashSerialized(obj: AnyRef) = Kryo.serializeAnyRef(obj).sha256
-  def hashSerializedBytes(obj: AnyRef) = Kryo.serializeAnyRef(obj).sha256Bytes
+  def hashSerialized(obj: AnyRef, withRefs: Boolean = false) = Kryo.serializeAnyRef(obj, withRefs).sha256
+
+  def hashSerializedBytes(obj: AnyRef, withRefs: Boolean = false) = Kryo.serializeAnyRef(obj, withRefs).sha256Bytes
 
   def signHashWithKey(hash: String, privateKey: PrivateKey): String =
     bytes2hex(signData(hash.getBytes())(privateKey))